requirements.txt
djangorestframework==3.16.1
psycopg==3.2.13
pycountry==24.6.1